Update 1.2.0
- seperate calendar components (normal, limited)
- change directory structure
- add public modifier to the calendar style properties (refs #15)
- update README
