+++ Example/YYCalendar/Base.lproj/Main.storyboard
@@ -27,7 +27,7 @@
                                 <constraints>
                                     <constraint firstAttribute="height" constant="40" id="dAw-Pi-aab"/>
                                 </constraints>
-                                <state key="normal" title="Show YYCalendar">
+                                <state key="normal" title="Show Normal Calendar">
                                     <color key="titleColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 </state>
                                 <connections>
@@ -52,39 +52,23 @@
                                 <constraints>
                                     <constraint firstAttribute="height" constant="40" id="Did-Af-aLK"/>
                                 </constraints>
-                                <state key="normal" title="Show YYCalendar (disable after today)">
+                                <state key="normal" title="Show Limited Calendar">
                                     <color key="titleColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 </state>
                                 <connections>
                                     <action selector="showCalendar2:" destination="vXZ-lx-hvc" eventType="touchUpInside" id="Ent-1k-B52"/>
                                 </connections>
                             </button>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Itb-Cl-UwP">
-                                <rect key="frame" x="46" y="332" width="283" height="40"/>
-                                <color key="backgroundColor" red="0.0" green="0.58980089430000004" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="40" id="xT7-Mb-u97"/>
-                                </constraints>
-                                <state key="normal" title="Show YYCalendar (limited period)">
-                                    <color key="titleColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                </state>
-                                <connections>
-                                    <action selector="showCalendar3:" destination="vXZ-lx-hvc" eventType="touchUpInside" id="SO4-Nm-2ei"/>
-                                </connections>
-                            </button>
                         </subviews>
                         <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                         <constraints>
                             <constraint firstAttribute="trailingMargin" secondItem="Q6e-Su-w08" secondAttribute="trailing" constant="30" id="1h4-ti-2ZF"/>
-                            <constraint firstAttribute="trailingMargin" secondItem="Itb-Cl-UwP" secondAttribute="trailing" constant="30" id="A1c-JA-bQn"/>
                             <constraint firstItem="Xon-wd-1Pm" firstAttribute="centerX" secondItem="kh9-bI-dsS" secondAttribute="centerX" id="Fbm-NH-u4g"/>
                             <constraint firstItem="zaX-cJ-bCm" firstAttribute="centerX" secondItem="kh9-bI-dsS" secondAttribute="centerX" id="PP5-gl-iTI"/>
                             <constraint firstItem="Q6e-Su-w08" firstAttribute="leading" secondItem="kh9-bI-dsS" secondAttribute="leadingMargin" constant="30" id="TTn-ZJ-2AC"/>
                             <constraint firstItem="zaX-cJ-bCm" firstAttribute="top" secondItem="Xon-wd-1Pm" secondAttribute="bottom" constant="30" id="h1A-kY-C7T"/>
                             <constraint firstAttribute="trailingMargin" secondItem="avE-zr-6kq" secondAttribute="trailing" constant="30" id="iYP-Yd-9ly"/>
-                            <constraint firstItem="Itb-Cl-UwP" firstAttribute="leading" secondItem="kh9-bI-dsS" secondAttribute="leadingMargin" constant="30" id="oCc-2o-Uzh"/>
                             <constraint firstItem="Xon-wd-1Pm" firstAttribute="top" secondItem="jyV-Pf-zRb" secondAttribute="bottom" constant="50" id="oYP-a8-z7j"/>
-                            <constraint firstItem="Itb-Cl-UwP" firstAttribute="top" secondItem="Q6e-Su-w08" secondAttribute="bottom" constant="30" id="qDX-vN-b0i"/>
                             <constraint firstItem="avE-zr-6kq" firstAttribute="top" secondItem="zaX-cJ-bCm" secondAttribute="bottom" constant="50" id="tI8-S5-REE"/>
                             <constraint firstItem="Q6e-Su-w08" firstAttribute="top" secondItem="avE-zr-6kq" secondAttribute="bottom" constant="30" id="uQh-Cu-E3m"/>
                             <constraint firstItem="avE-zr-6kq" firstAttribute="leading" secondItem="kh9-bI-dsS" secondAttribute="leadingMargin" constant="30" id="z7N-1W-kFL"/>
