+++ YYCalendar.podspec
@@ -8,7 +8,7 @@
 
 Pod::Spec.new do |s|
 	s.name             = 'YYCalendar'
-	s.version          = '1.1.0'
+	s.version          = '1.2.0'
 	s.swift_version    = '5.0'
 	s.summary          = 'Very Simple and Useful Calendar'
